patch 1.3
Done DiscoverScreen but haven't implemented OnTouch effects to Camera and Photo Mode
